+++ src/GraphUtil.ts
@@ -4,7 +4,7 @@ import * as j from './util/json'
 import * as caster from './util/GetExplicitGraph'
 import * as cloner from './util/GraphClone'
 import * as functionalFn from './util/functional'
-import * as flowAlgo from './util/MinMaxFlow'
+import * as flowAlgo from './util/MaxFlow'
 
 export const hasCycle = HasCycle.hasCycle
 
@@ -23,7 +23,4 @@ export const functional = {
   ...functionalFn
 }
 
-export const flow = {
-  max: flowAlgo.findMaxFlow,
-  min: flowAlgo.findMinFlow
-}
+export const findMaxFlow = flowAlgo.findMaxFlow
